#66 | Rewrote the ticker to tick the whole clock object
